web: improve query for getting volume templates
- Now the list of volumes is properly cached.
