+++ web/src/api/storage.ts
@@ -23,6 +23,8 @@
 import { get, post, put } from "~/api/http";
 import { Job } from "~/types/job";
 import { Action, config, configModel, ProductParams, Volume } from "~/api/storage/types";
+import { fetchDevices } from "~/api/storage/devices";
+import { StorageDevice, Volume as StorageVolume, VolumeTarget } from "~/types/storage";
 
 /**
  * Starts the storage probing process.
@@ -52,6 +54,33 @@ const fetchDefaultVolume = (mountPath: string): Promise<Volume | undefined> => {
   return get(`/api/storage/product/volume_for?mount_path=${path}`);
 };
 
+const fetchVolumeTemplates = async (): Promise<StorageVolume[]> => {
+  const buildVolume = (
+    rawVolume: Volume,
+    devices: StorageDevice[],
+    productMountPoints: string[],
+  ): StorageVolume => ({
+    ...rawVolume,
+    outline: {
+      ...rawVolume.outline,
+      // Indicate whether a volume is defined by the product.
+      productDefined: productMountPoints.includes(rawVolume.mountPath),
+    },
+    minSize: rawVolume.minSize || 0,
+    transactional: rawVolume.transactional || false,
+    target: rawVolume.target as VolumeTarget,
+    targetDevice: devices.find((d) => d.name === rawVolume.targetDevice),
+  });
+
+  const systemDevices = await fetchDevices("system");
+  const product = await fetchProductParams();
+  const mountPoints = ["", ...product.mountPoints];
+  const rawVolumes = await Promise.all(mountPoints.map(fetchDefaultVolume));
+  return rawVolumes
+    .filter((v) => v !== undefined)
+    .map((v) => buildVolume(v, systemDevices, product.mountPoints));
+};
+
 const fetchActions = (): Promise<Action[]> => get("/api/storage/devices/actions");
 
 /**
@@ -75,6 +104,7 @@ export {
   fetchUsableDevices,
   fetchProductParams,
   fetchDefaultVolume,
+  fetchVolumeTemplates,
   fetchActions,
   fetchStorageJobs,
   findStorageJob,
